rag_system/requirements.txt
# GEPA Framework - Core Package for Prompt Optimization
gepa

# LLM Integration (supports multiple API providers)
litellm>=1.0.0
groq>=0.4.0

# Vector Store - Using older stable version
chromadb==0.4.22

# Document Processing for Clinical Data
pandas>=2.0.0
openpyxl>=3.1.0

# Data Processing
numpy>=1.24.0

# Web Interface
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0
httpx>=0.24.0
